misc/system.c: cast size_t to long before using %ld
git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/branches/openvas-libraries-7.0@20404 423fd1db-d629-0410-8442-d21db03e70f4

+++ misc/system.c
@@ -125,7 +125,8 @@ erealloc (void *ptr, size_t size)
 
   if ((int) size < 0)
     {
-      log_legacy_write ("Won't realloc() a pointer of size %ld !", size);
+      log_legacy_write ("Won't realloc() a pointer of size %ld !",
+                        (long) size);
       exit (1);
     }
